requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
easyocr==1.7.1
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-documentai==2.31.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gspread==6.1.2
gunicorn==23.0.0
h11==0.14.0
idna==3.7
imageio==2.35.1
inference-sdk==0.17.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.4
supervision==0.23.0
sympy==1.13.2
thop==0.1.1.post2209072238
tifffile==2024.8.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.4
yarl==1.9.5
